refactor(electron-backend): split DB event handlers into modules
